Extracted the selection dialog to a new class InstantiateSelectionDialog.  Fixed up the labels on the selection list again.
